Changed models structure

+++ HoPet/Models/Product.cs
@@ -9,7 +9,7 @@ namespace HoPet.Models
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int Price { get; set; }
+        public double Price { get; set; }
         public int Quantity { get; set; }
         public AnimalType PetRelated { get; set; }
         public string Description { get; set; }
